aisih code modify

+++ simulator/ns-3.39/src/point-to-point/model/switch-mmu.h
@@ -19,7 +19,7 @@ public:
 	SwitchMmu(void);
 
 	//zqadd
-	uint64_t GetGHeadroom(uint32_t port, uint32_t qIndex,int index);
+	double GetGHeadroom(uint32_t port, uint32_t qIndex,int index);
 	uint64_t GetAIHeadroom();
 	void UpdateHeadroom(uint32_t port, uint32_t qIndex);
 	int64_t GetNowTime(); //ms
